Add OrdersSummaryExtractor

+++ src/main/java/com/silver/bars/liveorderboard/domain/Order.java
@@ -4,7 +4,6 @@ import lombok.Builder;
 import lombok.Data;
 
 import javax.validation.constraints.NotNull;
-import java.math.BigDecimal;
 
 @Data
 @Builder
@@ -13,7 +12,7 @@ public class Order {
     @NotNull
     private String userId;
     @NotNull
-    private BigDecimal quantity;
+    private Double quantity;
     @NotNull
     private Long price;
     @NotNull
